API RENAME: some function renamings, to make way for future move to doubles, or at least way from float-specificity: tenEigensolve  --> tenEigensolve_f tenMakeOne --> tenMakeOne_f tenSimulateOne --> tenSimulateOne_f tenAnisoCalc --> tenAnisoCalc_f tenEvqOne --> tenEvqOne_f tenEstimateLinearSingle --> tenEstimateLinearSingle_f API NEW: tenEigensolve_d API CHANGE: Functions xxx which haven't been renamed to xxx_f, and which took float arguments (not a pointer to float), now take a double instead. This is not a change that anyone will notice.  Affected functions are: tenEstimateLinear3D() tenEstimateLinear4D() tenExpand() tenSimulate() tenAnisoVolume() tenSizeScale() tenAnisoScale() tenEigenvaluePower() tenEigenvalueClamp() tenEigenvalueAdd()
API CHANGE:
More of a big deal: this guy took a float array ("weight"), and
  now its a double array:
tenSizeNormalize(Nrrd *nout, Nrrd *nin, double weight[3],
                 double amount, double target);


git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2190 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/ten/epireg.c
@@ -190,7 +190,7 @@ _tenEpiRegBlur(Nrrd **nblur, Nrrd **ndwi, int dwiLen,
 }
 
 int
-_tenEpiRegFindThresh(float *DWthrP, Nrrd **nin, int ninLen, int save) {
+_tenEpiRegFindThresh(double *DWthrP, Nrrd **nin, int ninLen, int save) {
   char me[]="_tenEpiRegFindThresh", err[AIR_STRLEN_MED];
   Nrrd *nhist, *ntmp;
   airArray *mop;
@@ -246,7 +246,7 @@ _tenEpiRegFindThresh(float *DWthrP, Nrrd **nin, int ninLen, int save) {
 
 int
 _tenEpiRegThreshold(Nrrd **nthresh, Nrrd **nblur, int ninLen,
-		    float DWthr, int verb, int progress) {
+		    double DWthr, int verb, int progress) {
   char me[]="_tenEpiRegThreshold", err[AIR_STRLEN_MED];
   airArray *mop;
   int I, sx, sy, sz, ni;
@@ -998,8 +998,8 @@ _tenEpiRegWarp(Nrrd **ndone, Nrrd *npxfr, Nrrd *nhst, Nrrd *ngrad,
 int
 tenEpiRegister3D(Nrrd **nout, Nrrd **nin, int ninLen, Nrrd *_ngrad,
 		 int reference,
-		 float bwX, float bwY, float fitFrac,
-		 float DWthr, int doCC, 
+		 double bwX, double bwY, double fitFrac,
+		 double DWthr, int doCC, 
 		 const NrrdKernel *kern, double *kparm,
 		 int progress, int verbose) {
   char me[]="tenEpiRegister3D", err[AIR_STRLEN_MED];
@@ -1134,8 +1134,8 @@ tenEpiRegister3D(Nrrd **nout, Nrrd **nin, int ninLen, Nrrd *_ngrad,
 int
 tenEpiRegister4D(Nrrd *_nout, Nrrd *_nin, Nrrd *ngrad,
 		 int reference,
-		 float bwX, float bwY, float fitFrac,
-		 float DWthr, int doCC, 
+		 double bwX, double bwY, double fitFrac,
+		 double DWthr, int doCC, 
 		 const NrrdKernel *kern, double *kparm,
 		 int progress, int verbose) {
   char me[]="tenEpiRegister4D", err[AIR_STRLEN_MED];
